entities configuration

+++ iSpend.Domain/Entities/Expense.cs
@@ -6,22 +6,22 @@ public sealed class Expense : Entity
 {
     public Guid UserId { get; private set; }
     public string Name { get; private set; }
-    public float Value { get; private set; }
+    public decimal Value { get; private set; }
     public bool Active { get; private set; }
     public int Duration { get; private set; }
     public int PaidMonths { get; private set; }
 
-    public Expense(string name, float value, bool active, int duration, int paidMonths, string userId)
+    public Expense(string name, decimal value, bool active, int duration, int paidMonths, string userId)
     {
         ValidateDomain(name, value, active, duration, paidMonths, userId);
     }
 
-    public void Update(string name, float value, bool active, int duration, int paidMonths)
+    public void Update(string name, decimal value, bool active, int duration, int paidMonths)
     {
         ValidateDomain(name, value, active, duration, paidMonths, this.UserId.ToString(), this.RegisteredAt);
     }
 
-    private void ValidateDomain(string name, float value, bool active, int duration, int paidMonths, string userId, DateTime? registeredAt = null)
+    private void ValidateDomain(string name, decimal value, bool active, int duration, int paidMonths, string userId, DateTime? registeredAt = null)
     {
         Guid validGuid;
         DomainExceptionValidation.When(string.IsNullOrEmpty(name), "Invalid name. Name is required");
