+++ iSpend.Domain/Entities/Income.cs
@@ -7,20 +7,20 @@ public sealed class Income : Entity
     public Guid UserId { get; private set; }
     public string Name { get; private set; }
     public bool Recurrent { get; private set; }
-    public float Value { get; private set; }
+    public decimal Value { get; private set; }
     public bool Active { get; private set; }
 
-    public Income(string name, float value, bool active, bool recurrent, string userId)
+    public Income(string name, decimal value, bool active, bool recurrent, string userId)
     {
         ValidateDomain(name, value, active, recurrent, userId);
     }
 
-    public void Update(string name, float value, bool active, bool recurrent)
+    public void Update(string name, decimal value, bool active, bool recurrent)
     {
         ValidateDomain(name, value, active, recurrent, this.UserId.ToString(), this.RegisteredAt);
     }
 
-    private void ValidateDomain(string name, float value, bool active, bool recurrent, string userId, DateTime? registeredAt = null)
+    private void ValidateDomain(string name, decimal value, bool active, bool recurrent, string userId, DateTime? registeredAt = null)
     {
         Guid validGuid;
         DomainExceptionValidation.When(string.IsNullOrEmpty(name), "Invalid name. Name is required");
