+++ iSpend.Domain/Entities/Subscription.cs
@@ -5,15 +5,15 @@ namespace iSpend.Domain.Entities;
 public sealed class Subscription : Entity
 {
     public string Name { get; private set; }
-    public float Price { get; private set; }
+    public decimal Price { get; private set; }
     public bool Active { get; private set; }
 
-    public Subscription(string name, float price, bool active, int creditCardId)
+    public Subscription(string name, decimal price, bool active, int creditCardId)
     {
         ValidateDomain(name, price, active, creditCardId);
     }
 
-    public void Update(string name, float price, bool active, int creditCardId)
+    public void Update(string name, decimal price, bool active, int creditCardId)
     {
         ValidateDomain(name, price, active, creditCardId, this.RegisteredAt);
     }
@@ -21,7 +21,7 @@ public sealed class Subscription : Entity
     public int CreditCardId { get; set; }
     public CreditCard CreditCard { get; set; }
 
-    private void ValidateDomain(string name, float price, bool active, int creditCardId, DateTime? registeredAt = null)
+    private void ValidateDomain(string name, decimal price, bool active, int creditCardId, DateTime? registeredAt = null)
     {
         DomainExceptionValidation.When(string.IsNullOrEmpty(name), "Invalid name. Name is required");
         DomainExceptionValidation.When((price < 0), "Invalid price. Price cannot be less than 0");
